shit

+++ back_end/restservice/src/main/java/com/goj/restservice/repository/SourceCodeRepository.java
@@ -7,6 +7,6 @@ import com.goj.restservice.entity.SourceCode;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface SourceCodeRepository extends PagingAndSortingRepository<SourceCode, Integer> {
+public interface SourceCodeRepository extends PagingAndSortingRepository<SourceCode, Long> {
 
 }
